+++ src/Serilog.Sinks.ApplicationInsights.Tests/CustomTelemetryConversionTest.cs
@@ -1,13 +1,10 @@
 ﻿using Microsoft.ApplicationInsights.Channel;
 using Microsoft.ApplicationInsights.DataContracts;
-using Microsoft.ApplicationInsights.Extensibility;
 using Serilog.Events;
-using Serilog.ExtensionMethods;
 using Serilog.Sinks.ApplicationInsights.Sinks.ApplicationInsights.TelemetryConverters;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Xunit;
 
 namespace Serilog.Sinks.ApplicationInsights.Tests
@@ -27,68 +24,44 @@ namespace Serilog.Sinks.ApplicationInsights.Tests
             Assert.Equal("test", LastSubmittedTraceTelemetry.Message);
         }
 
-        private class CustomConverter : ITelemetryConverter
+        private class CustomConverter : TraceTelemetryConverter
         {
-            private static ITelemetry GetTelemetry(LogEvent logEvent, IFormatProvider formatProvider)
-            {
-                if (logEvent.Exception != null)
-                {
-                    // Exception telemetry
-                    return logEvent.ToDefaultExceptionTelemetry(
-                    formatProvider,
-                    includeLogLevelAsProperty: false,
-                    includeRenderedMessageAsProperty: false,
-                    includeMessageTemplateAsProperty: false);
-                }
-                else
-                {
-                    // default telemetry
-                    return logEvent.ToDefaultTraceTelemetry(
-                    formatProvider,
-                    includeLogLevelAsProperty: false,
-                    includeRenderedMessageAsProperty: false,
-                    includeMessageTemplateAsProperty: false);
-                }
-            }
-
-            public IEnumerable<ITelemetry> Convert(LogEvent logEvent, IFormatProvider formatProvider)
+            public override IEnumerable<ITelemetry> Convert(LogEvent logEvent, IFormatProvider formatProvider)
             {
                 // first create a default TraceTelemetry using the sink's default logic
                 // .. but without the log level, and (rendered) message (template) included in the Properties
-                var telemetry = GetTelemetry(logEvent, formatProvider);
-
-                // then go ahead and post-process the telemetry's context to contain the user id as desired
-                if (logEvent.Properties.ContainsKey("UserId"))
+                foreach (ITelemetry telemetry in base.Convert(logEvent, formatProvider))
                 {
-                    telemetry.Context.User.Id = logEvent.Properties["UserId"].ToString();
+                    // then go ahead and post-process the telemetry's context to contain the user id as desired
+                    if (logEvent.Properties.ContainsKey("UserId"))
+                    {
+                        telemetry.Context.User.Id = logEvent.Properties["UserId"].ToString();
+                    }
+                    // post-process the telemetry's context to contain the operation id
+                    if (logEvent.Properties.ContainsKey("operation_Id"))
+                    {
+                        telemetry.Context.Operation.Id = logEvent.Properties["operation_Id"].ToString();
+                    }
+                    // post-process the telemetry's context to contain the operation parent id
+                    if (logEvent.Properties.ContainsKey("operation_parentId"))
+                    {
+                        telemetry.Context.Operation.ParentId = logEvent.Properties["operation_parentId"].ToString();
+                    }
+                    // typecast to ISupportProperties so you can manipulate the properties as desired
+                    ISupportProperties propTelematry = (ISupportProperties)telemetry;
+
+                    // find redundent properties
+                    var removeProps = new[] { "UserId", "operation_parentId", "operation_Id" };
+                    removeProps = removeProps.Where(prop => propTelematry.Properties.ContainsKey(prop)).ToArray();
+
+                    foreach (var prop in removeProps)
+                    {
+                        // remove redundent properties
+                        propTelematry.Properties.Remove(prop);
+                    }
+
+                    yield return telemetry;
                 }
-                // post-process the telemetry's context to contain the operation id
-                if (logEvent.Properties.ContainsKey("operation_Id"))
-                {
-                    telemetry.Context.Operation.Id = logEvent.Properties["operation_Id"].ToString();
-                }
-                // post-process the telemetry's context to contain the operation parent id
-                if (logEvent.Properties.ContainsKey("operation_parentId"))
-                {
-                    telemetry.Context.Operation.ParentId = logEvent.Properties["operation_parentId"].ToString();
-                }
-                // typecast to ISupportProperties so you can manipulate the properties as desired
-                ISupportProperties propTelematry = (ISupportProperties)telemetry;
-
-                // find redundent properties
-                var removeProps = new[] { "UserId", "operation_parentId", "operation_Id" };
-                removeProps = removeProps.Where(prop => propTelematry.Properties.ContainsKey(prop)).ToArray();
-
-                foreach (var prop in removeProps)
-                {
-                    // remove redundent properties
-                    propTelematry.Properties.Remove(prop);
-                }
-
-                var tpcb = TelemetryConfiguration.Active.TelemetryProcessorChainBuilder;
-
-
-                yield return telemetry;
             }
         }
     }
