+++ src/Serilog.Sinks.ApplicationInsights/Sinks/ApplicationInsights/ApplicationInsightsSink.cs
@@ -12,54 +12,199 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 using System;
 using System.Collections.Generic;
+using System.Reflection;
+using System.Runtime.ExceptionServices;
+using System.Threading;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Channel;
+using Serilog.Core;
 using Serilog.Events;
+using Serilog.Sinks.ApplicationInsights.Sinks.ApplicationInsights.TelemetryConverters;
 
 namespace Serilog.Sinks.ApplicationInsights
 {
     /// <summary>
-    /// Writes log events to a Microsoft Azure Application Insights account.
+    /// Base class for Microsoft Azure Application Insights based Sinks.
+    /// Inspired by their NLog Appender implementation.
     /// </summary>
-    public class ApplicationInsightsSink : ApplicationInsightsSinkBase
+    class ApplicationInsightsSink : ILogEventSink, IDisposable
     {
+        private long _isDisposing = 0;
+        private long _isDisposed = 0;
+
+        private TelemetryClient _telemetryClient;
+
         /// <summary>
-        /// Creates a sink that saves logs as <see cref="ITelemetry"/> to the Application Insights account for the given <paramref name="telemetryClient" /> instance.
+        /// Gets or sets a value indicating whether this instance is being disposed.
         /// </summary>
-        /// <param name="telemetryClient">Required Application Insights <paramref name="telemetryClient" />.</param>
-        /// <param name="formatProvider">Supplies culture-specific formatting information, or null for default provider.</param>
-        /// <param name="logEventToTelemetryConverter">The <see cref="LogEvent" /> to <see cref="ITelemetry" /> converter.</param>
-        /// <exception cref="ArgumentNullException"><paramref name="telemetryClient" /> is <see langword="null" />.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="logEventToTelemetryConverter" /> is <see langword="null" />.</exception>
-        public ApplicationInsightsSink(
-            TelemetryClient telemetryClient,
-            Func<LogEvent, IFormatProvider, ITelemetry> logEventToTelemetryConverter,
-            IFormatProvider formatProvider = null)
-            : base(telemetryClient, (e, f, c) => new[] { logEventToTelemetryConverter(e, f) }, formatProvider)
+        /// <value>
+        /// <c>true</c> if this instance is being disposed; otherwise, <c>false</c>.
+        /// </value>
+        public bool IsDisposing
         {
-            if (telemetryClient == null) throw new ArgumentNullException(nameof(telemetryClient));
-            if (logEventToTelemetryConverter == null) throw new ArgumentNullException(nameof(logEventToTelemetryConverter));
+            get => Interlocked.Read(ref _isDisposing) == 1;
+            protected set => Interlocked.Exchange(ref _isDisposing, value ? 1 : 0);
         }
 
         /// <summary>
-        /// Creates a sink that saves logs as <see cref="ITelemetry"/> to the Application Insights account for the given <paramref name="telemetryClient" /> instance.
+        /// Gets a value indicating whether this instance has been disposed.
+        /// </summary>
+        /// <value>
+        /// <c>true</c> if this instance has been disposed; otherwise, <c>false</c>.
+        /// </value>
+        public bool IsDisposed
+        {
+            get => Interlocked.Read(ref _isDisposed) == 1;
+            protected set => Interlocked.Exchange(ref _isDisposed, value ? 1 : 0);
+        }
+
+        private readonly IFormatProvider _formatProvider;
+
+        private readonly ITelemetryConverter _telemetryConverter;
+
+        /// <summary>
+        /// Creates a sink that saves logs to the Application Insights account for the given <paramref name="telemetryClient" /> instance.
         /// </summary>
         /// <param name="telemetryClient">Required Application Insights <paramref name="telemetryClient" />.</param>
+        /// <param name="telemetryConverter">The <see cref="LogEvent"/> to <see cref="ITelemetry"/> converter.</param>
         /// <param name="formatProvider">Supplies culture-specific formatting information, or null for default provider.</param>
-        /// <param name="logEventToTelemetryConverter">The <see cref="LogEvent" /> to <see cref="ITelemetry" /> converter.</param>
-        /// <exception cref="ArgumentNullException"><paramref name="telemetryClient" /> is <see langword="null" />.</exception>
-        /// <exception cref="ArgumentNullException"><paramref name="logEventToTelemetryConverter" /> is <see langword="null" />.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="telemetryClient" /> cannot be null</exception>
         public ApplicationInsightsSink(
             TelemetryClient telemetryClient,
-            Func<LogEvent, IFormatProvider, TelemetryClient, IEnumerable<ITelemetry>> logEventToTelemetryConverter,
+            ITelemetryConverter telemetryConverter,
             IFormatProvider formatProvider = null)
-            : base(telemetryClient, logEventToTelemetryConverter, formatProvider)
         {
-            if (telemetryClient == null) throw new ArgumentNullException(nameof(telemetryClient));
-            if (logEventToTelemetryConverter == null) throw new ArgumentNullException(nameof(logEventToTelemetryConverter));
+            _telemetryClient = telemetryClient ?? throw new ArgumentNullException(nameof(telemetryClient));
+            _telemetryConverter = telemetryConverter ?? throw new ArgumentNullException(nameof(telemetryConverter));
+
+            _formatProvider = formatProvider;
         }
+
+        #region AI specifc Helper methods
+
+        /// <summary>
+        /// Hands over the <paramref name="telemetry" /> to the AI telemetry client.
+        /// </summary>
+        /// <param name="telemetry">The telemetry.</param>
+        /// <exception cref="System.ArgumentNullException">telemetry</exception>
+        protected virtual void TrackTelemetry(ITelemetry telemetry)
+        {
+            if (telemetry == null) throw new ArgumentNullException(nameof(telemetry));
+
+            CheckForAndThrowIfDisposed();
+
+            // the .Track() method is save to use (even though documented otherwise)
+            // see https://github.com/Microsoft/ApplicationInsights-dotnet/issues/244
+            _telemetryClient?.Track(telemetry);
+        }
+
+        #endregion AI specifc Helper methods
+
+        #region Implementation of ILogEventSink
+
+        /// <summary>
+        /// Emit the provided log event to the sink.
+        /// </summary>
+        /// <param name="logEvent">The log event to write.</param>
+        /// <exception cref="Exception">A delegate callback throws an exception.</exception>
+        /// <exception cref="TargetInvocationException">A delegate callback throws an exception.</exception>
+        public virtual void Emit(LogEvent logEvent)
+        {
+            if (logEvent == null) throw new ArgumentNullException(nameof(logEvent));
+
+            CheckForAndThrowIfDisposed();
+
+            try
+            {
+                IEnumerable<ITelemetry> telemetries = _telemetryConverter.Convert(logEvent, _formatProvider);
+
+                // if 'null' is returned (& we therefore there's nothing to track), the logEvent is basically skipped
+                if (telemetries != null)
+                {
+                    foreach (ITelemetry telemetry in telemetries)
+                    {
+                        if (telemetry != null)
+                        {
+                            TrackTelemetry(telemetry);
+                        }
+                    }
+                }
+            }
+            catch (TargetInvocationException targetInvocationException)
+            {
+                // rethrow original exception (inside the TargetInvocationException) if any
+                if (targetInvocationException.InnerException != null)
+                {
+                    ExceptionDispatchInfo.Capture(targetInvocationException.InnerException).Throw();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+        }
+
+        #endregion
+
+        #region Implementation of IDisposable
+
+        /// <summary>
+        /// Checks whether this instance has been disposed and if so, throws an <see cref="ObjectDisposedException"/>.
+        /// </summary>
+        /// <exception cref="ObjectDisposedException"></exception>
+        protected void CheckForAndThrowIfDisposed()
+        {
+            if (IsDisposed)
+            {
+                throw new ObjectDisposedException(this.GetType().Name);
+            }
+        }
+
+        /// <summary>
+        /// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
+        /// </summary>
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        /// <summary>
+        /// Releases unmanaged and - optionally - managed resources.
+        /// </summary>
+        /// <param name="disposeManagedResources"><c>true</c> to release both managed and unmanaged resources; <c>false</c> to release only unmanaged resources.</param>
+        protected virtual void Dispose(bool disposeManagedResources)
+        {
+            if (IsDisposing || IsDisposed)
+                return;
+
+            try
+            {
+                IsDisposing = true;
+
+                // we only have managed resources to dispose of
+                if (disposeManagedResources)
+                {
+                    // attempt to free managed resources
+                    try
+                    {
+                        _telemetryClient?.Flush();
+                    }
+                    finally
+                    {
+                        _telemetryClient = null;
+                    }
+                }
+            }
+            finally
+            {
+                IsDisposed = true;
+                IsDisposing = false;
+            }
+        }
+
+        #endregion Implementation of IDisposable
     }
 }
