+++ src/Serilog.Sinks.ApplicationInsights/Sinks/ApplicationInsights/Formatters/IValueFormatter.cs
@@ -1,13 +1,9 @@
 ﻿using Serilog.Events;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Serilog.Sinks.ApplicationInsights.Sinks.ApplicationInsights.Formatters
 {
-    interface IValueFormatter
+    public interface IValueFormatter
     {
         void Format(string propertyName, LogEventPropertyValue propertyValue, IDictionary<string, string> properties);
     }
