+++ src/Serilog.Sinks.ApplicationInsights/Sinks/ApplicationInsights/TelemetryConverters/EventTelemetryConverter.cs
@@ -6,21 +6,16 @@ using Serilog.Events;
 
 namespace Serilog.Sinks.ApplicationInsights.Sinks.ApplicationInsights.TelemetryConverters
 {
-    public class TraceTelemetryConverter : TelemetryConverterBase
+    public class EventTelemetryConverter : TelemetryConverterBase
     {
         public override IEnumerable<ITelemetry> Convert(LogEvent logEvent, IFormatProvider formatProvider)
         {
-            if (logEvent == null)
-                throw new ArgumentNullException(nameof(logEvent));
+            if (logEvent == null) throw new ArgumentNullException(nameof(logEvent));
 
             if (logEvent.Exception == null)
             {
-                var renderedMessage = logEvent.RenderMessage(formatProvider);
-
-                var telemetry = new TraceTelemetry(renderedMessage)
-                {
-                    Timestamp = logEvent.Timestamp,
-                    SeverityLevel = ToSeverityLevel(logEvent.Level)
+                var telemetry = new EventTelemetry(logEvent.MessageTemplate.Text) {
+                    Timestamp = logEvent.Timestamp
                 };
 
                 // write logEvent's .Properties to the AI one
